requirements.txt
streamlit==1.32.2
pypdf==4.1.0
numpy==1.26.4
pandas==2.2.1
faiss-cpu==1.8.0
sentence-transformers==2.6.1
python-dotenv==1.0.1
google-generativeai==0.5.4
ddgs==9.5.2